API

Decorator macros for loops

Checkpointing.@ad_checkpoint — Macro

@ad_checkpoint(
    alg,
    loop,
)

This macro is supposed to be only used in conjunction with EnzymeRules. It does not initialize the shadowcopy. Apply the checkpointing scheme alg on the loop loop expression.

Supported Schemes

Checkpointing.Revolve — Type

Revolve{MT}(
    checkpoints::Int;
    storage::AbstractStorage = ArrayStorage{MT}(checkpoints),
    verbose::Int = 0,
    gc::Bool = true,
    write_checkpoints::Bool = false,
    write_checkpoints_period::Int = 1,
    write_checkpoints_filename::String = "chkp",
) where {MT}

Creates a new Revolve object for checkpointing.

checkpoints: is the number of checkpoints used for storage.

storage: is the storage backend to use (default is ArrayStorage).

verbose::Int: Verbosity level for logging and diagnostics.

gc::Bool: Whether to enable garbage collection (default is true).

write_checkpoints::Bool: Whether to enable writing checkpoints (default is false).

write_checkpoints_period::Int: The period for writing checkpoints (default is 1).

write_checkpoints_filename::String: The filename for writing checkpoints (default is "chkp").

References

Griewank, A. & Walther, A. “Algorithm 799: Revolve: An Implementation of

Checkpointing for the Reverse or Adjoint Mode of Computational Differentiation.” ACM Transactions on Mathematical Software.

This documentation outlines the structure, usage, and functionality of Revolve and should help users integrate the checkpointing scheme into their Julia projects.

Checkpointing.Periodic — Type

Periodic(
    checkpoints::Int;
    storage::AbstractStorage = ArrayStorage{MT}(checkpoints),
    verbose::Int = 0,
    gc::Bool = true,
    write_checkpoints::Bool = false,
) where {MT}

The periodic scheme is used to store the state of the system at regular intervals and then restore it when needed.

checkpoints: is the number of checkpoints used for storage.

storage: is the storage backend to use (default is ArrayStorage).

verbose::Int: Verbosity level for logging and diagnostics.

gc::Bool: Whether to enable garbage collection (default is true).

write_checkpoints::Bool: Whether to enable writing checkpoints (default is false).

The period will be div(steps, checkpoints).

Checkpointing.Online_r2 — Type

Online_r2{FT}(
    checkpoints::Int,
    storage::AbstractStorage = ArrayStorage{FT}(checkpoints);
    verbose::Int = 0,
) where {FT}

This implements the online checkpointing scheme for r=2 as described in the paper "New Algorithms for Optimal Online Checkpointing" by Philipp Stumm and Andrea Walther.

The key advantage is that the number of iterations does not have to be known a priori.

Creates a new Online_r2 object for checkpointing.

checkpoints: is the number of checkpoints used for storage.

storage: is the storage backend to use (default is ArrayStorage).

Supported Storages

Checkpointing.ArrayStorage — Type

ArrayStorage

Array (RAM) storage for checkpointing.

Checkpointing.HDF5Storage — Type

HDF5Storage

A storage type for checkpointing that uses HDF5 files to store the checkpoints.

Developer variables for implementing new schemes

Checkpointing.Scheme — Type

Scheme

Abstract type from which all checkpointing schemes are derived.